requirements.txt
openai>=1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytz
python-dateutil